Couple the AVR and TWI drivers more tightly.
This fixes a system lockup bug introduced by 5856ed87b0ed, due to the
spinlocking and complex notification mechanism built into the TWI
driver. That code was necessary on the assumption that something other
than the coprocessor was on the I2C bus, which is false. So, this makes
the code simpler, and forces the AVR and TWI drivers to work in
lockstep, eliminating the risk of locking up.