Prepped a radar demo.
First, we check for the radar presence. If the radar is not found, the
user can press OK to search it again. Otherwise, OK will switch to
measurement mode, where R0 is read every 500ms. The user can exit from
measurement mode and halt the system at any time using the CANCEL
button.