Rewrite the interrupt handler code to handle task saving.
When entered from a kernel mode, the handler behaves exactly as it
used to. However, if the handler is entered from either of user or
system mode, it saves all the user task's state in the task's own
stack. A future scheduler will then be able to swap in a different
"loaded" stack pointer, and as a result perform a task switch.