Add a radar test.
Provide a radar test for the NxOS test suite. First, the test tries
to detect the radar, and loops on detection until a radar is found
(the user must press "Ok" to detect again). When a radar has been
found, its information is displayed. After a press on "Ok", the
tests goes into radar scanning mode and displays the 8 radar readings
20 times. The test can be cancelled before the 20th iteration by
pressing "Right".