Add a fault detector in the baseplate main function.
This fault detector checks if the system is booting as a result
of a watchdog or brownout fault, and crashes with a message if so.

Brownout and watchdog faults are theoretically "can't happen" errors
for us, but they produce most confusing symptoms when they do manage to
happen (such as when this dummy incorrectly disables the watchdog timer).
It's not that much extra code, and better safe than sorry.