Change the system tick driver to have a dedicated scheduler interrupt.
All bookkeeping tasks are now done directly in the time-critical interrupt,
since everything is fairly short. If a scheduler callback is registered, a
lower priority interrupt is triggered to service it at a lower priority.

The scheduler runs at a priority lower than the system timer and TWI driver,
so that the AVR link can never be killed by the scheduler, but is higher than
other drivers, to give more importance to the scheduler.

If no scheduler callback is registered, that also means that now the system
timer IRQ no longer triggers a second IRQ. Nice.