Implement a demonstration of semaphores in Marvin's main().
This demo implements a producer/consumer beeping task pair. One task "produces"
beep tokens by incrementing a semaphore, and the other "consumes" the tokens
by emitting a single beep for each time it successfully decremented the
semaphore. The result is a rythmic 2s beep, but produced by the synchronized
cooperation of two tasks.