Allow to use toolchains without thumb-interwork.
That means we are searching libgcc.a in the subdirectories 'interwork',
'thumb' and '' and are using the corresponding CFLAGS and ASFLAGS.

Choosing which library is used is (currently) not possible, we are using
either thumb-interwork, thumb, or plain (searched in that order).

It is shown, which library will be used.

Signed-off-by: danderson <dave@natulte.net>